+++ core/src/server/Server.cpp
@@ -24,6 +24,7 @@
 #include "server/DBWrapper.h"
 #include "server/grpc_impl/GrpcServer.h"
 #include "server/init/CpuChecker.h"
+#include "server/init/GpuChecker.h"
 #include "server/web_impl/WebServer.h"
 #include "src/version.h"
 //#include "storage/s3/S3ClientWrapper.h"
@@ -237,6 +238,13 @@ Server::Start() {
         if (!s.ok()) {
             return s;
         }
+
+#ifdef MILVUS_GPU_VERSION
+        s = GpuChecker::CheckGpuEnvironment();
+        if (!s.ok()) {
+            return s;
+        }
+#endif
         /* record config and hardware information into log */
         LogConfigInFile(config_filename_);
         LogCpuInfo();