+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#2054 Check if CPU instruction sets are illegal
 -   \#2059 Add lock file avoid multiple instances modifying data at the same time
 -   \#2064 Warn when use SQLite as metadata management
+-   \#2111 Check GPU environment before start server
 
 ## Improvement
 -   \#221 Refactor LOG macro